Create genetic algorithm skeleton

+++ TravelingThiefProblem/src/pwr/jakubgruda/domain/City.kt
@@ -1,3 +1,3 @@
 package pwr.jakubgruda.domain
 
-data class City(val index: Int, val x: Int, val y: Int)
+data class City(val index: Int, val x: Double, val y: Double)
